+++ src/Services/wallet.service.ts
@@ -27,51 +27,76 @@ export class WalletService{
             return new Promise(e=>e({ success:true, value:wallet } as Result<Wallet>))
         }
     }
-    async FindOneForeignID(foreign_id:number):Promise<Result<Wallet>>{
-        let wallet:Wallet=(await this.db.get(`SELECT * FROM wallets WHERE foreign_id=${foreign_id}`)) as Wallet;
+    async FindOneForeignID(foreign_id:number,roleid:number):Promise<Result<Wallet>>{
+        let wallet:Wallet=(await this.db.get(`SELECT * FROM wallets WHERE foreign_id=${foreign_id} AND roleid=${roleid}`)) as Wallet; 
+
         if(typeof wallet == "undefined"){
             return new Promise(e=>e({success:false,value:{} as Wallet} as Result<Wallet>));
         }else{
             return new Promise(e=>e({success:true,value:wallet} as Result<Wallet>));
         }
     }
-    async FindManyForeignID(foreign_id:number):Promise<Result<Wallet[]>>{
-        let wallet:Wallet[]=(await this.db.all(`SELECT * FROM wallets WHERE foreign_id=${foreign_id}`)) as Wallet[];
-        if(typeof wallet =="undefined"){
-            return new Promise(e=>e({success:false,value:[] as Wallet[]}as Result<Wallet[]>));
-        }else{
-            return new Promise(e=>e({success:true,value:wallet} as Result<Wallet[]>));
-        }
-    }
-    async AddOne(foreign_id:number,amount:number):Promise<Result<Wallet>>{
-        await this.db.exec(`INSERT INTO wallets (foreign_id,amount) VALUES (${foreign_id},${amount});`);
-        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id);
-        if(JSON.stringify(wallet.value)!="{}" ){
+    async AddOne(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
+        await this.db.exec(`INSERT INTO wallets (foreign_id,roleid,amount) VALUES (${foreign_id},${roleid},${amount});`);
+        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
+        if(wallet.success){
             return new Promise(e=>e({success:true,value:wallet.value} as Result<Wallet>));
         }else{
             return new Promise(e=>e({success:false,value:{} as Wallet} as Result<Wallet>));
         }
     }
-    async DeleteOne(foreign_id:number):Promise<Result<Wallet>>{
-        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id);
+    async DeleteOne(foreign_id:number,roleid:number):Promise<Result<Wallet>>{
+        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
         if(JSON.stringify(wallet.value)!="{}"){
-            await this.db.exec(`DELETE FROM wallets WHERE foreign_id=${foreign_id}`);
+            await this.db.exec(`DELETE FROM wallets WHERE foreign_id=${foreign_id} AND roleid=${roleid}`);
             return new Promise(e=>e({success:true,value:wallet.value} as Result<Wallet>));
         }
         else{
             return new Promise(e=>e({success:false,value:{} as Wallet} as Result<Wallet>));
         }
     }
-    async UpdateOne(foreign_id:number,amount:number):Promise<Result<Wallet>>{
-        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id);
+    async UpdateOne(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
+        let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
         if(JSON.stringify(wallet.value)!="{}"){
-            await this.db.exec(`UPDATE wallets SET amount=${amount} WHERE foreign_id=${foreign_id};`);
-            let wallet_updated:Result<Wallet>=await this.FindOneForeignID(foreign_id);
+            await this.db.exec(`UPDATE wallets SET amount=${amount} WHERE foreign_id=${foreign_id} AND roleid=${roleid};`);
+            let wallet_updated:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
             return new Promise(e=>e({success:true,value:wallet_updated.value} as Result<Wallet>));
         }else{
             return new Promise(e=>e({success:false,value:{} as Wallet} as Result<Wallet>));
         }
-
     }
-
+    async AddMoney(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
+        let wallet=await this.FindOneForeignID(foreign_id,roleid);
+        if(wallet.success){
+            let amount_=wallet.value.amount;
+            amount_+=amount;
+            let updated:Result<Wallet>=await this.UpdateOne(foreign_id,roleid,amount_);
+            if(updated.success){
+                return new Promise(e=>e({success:true,value:updated.value}));
+            }else{
+                return new Promise(e=>e({success:false,value:{} as Wallet}));
+            }
+        }else{
+            return new Promise(e=>e({success:false,value:{} as Wallet}));
+        }
+    }
+    async WithdrawMoney(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
+        let wallet=await this.FindOneForeignID(foreign_id,roleid);
+        if(wallet.success){
+            let amount_=wallet.value.amount;
+            if(amount<=amount_){
+                amount_-=amount;
+                let updated:Result<Wallet>=await this.UpdateOne(foreign_id,roleid,amount_);
+                if(updated.success){
+                    return new Promise(e=>e({success:true,value:updated.value}));
+                }else{
+                    return new Promise(e=>e({success:false,value:{} as Wallet}));
+                }
+            }else{
+                return new Promise(e=>e({success:false,value:{} as Wallet}));
+            }
+        }else{
+            return new Promise(e=>e({success:false,value:{} as Wallet}));
+        }
+    }
 }
